refactor: 거리계산함수 재상용을 위한 분화
